feat: implement E2E tests for all bot handlers
- Refactor bot initialization into src/app.ts for testability
- Use real application wiring in E2E tests
- Add coverage for all bot commands and events
- Fix botInfo mock and TypeScript issues in tests
- Ensure E2E tests are correctly type-checked in tsconfig.json
