+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@gabrielrufino/pdf-studio-bot",
-  "version": "1.10.2",
+  "version": "1.10.3",
   "private": true,
   "packageManager": "pnpm@10.32.1",
   "author": "Gabriel Rufino <contato@gabrielrufino.com>",
